refactor: move InMemDatabase into iroh lib as well
tests/examples for iroh-bytes that might want to use this can live in iroh
